Connect steps one and two of IR pipeline
