+++ evaluation/batch_eval_rag.py
@@ -1,4 +1,4 @@
-from retriever import ImageRetrieval
+from ir.image_retrieval import ImageRetrieval
 
 from eval import ArtEvaluator
 from pathlib import Path
